Changing metadata to events array
Instead of sending an object with metadata, we'll just send a list of events. Any server-side module can add an event to a session, and the server will send the accumulated list of events with the next request. The client-side view module processes the list of events. Each event has a unique identifier and various properties depending on the identifier.