Add FFT data manager, improve UI
The FFT data manager keeps track of the array buffers that store FFT data. Since our two AnalyserNodes add delay to the audio graph (their data is 20k and 16k samples ahead of the audio that's being played), we need to cache their data and only show it once the audio is at that time.