🐎 Move tempo adjustment into separate thread
The tempo adjustment was blocking the server for a long time, especially when slowing down a song. So we need to move it to a separate thread so we can continue serving HTTP requests.
The current workaround is very hacky 💩👎 so eventually we'll need to switch from tiny-worker to a better library like Napa.js.